requirements-docker.txt
torch==2.5.1
torchvision==0.20.1
numpy==1.23.5
opencv-python==4.10.0.84
Pillow==8.4.0
tensorrt==10.6.0
packaging
transformers==4.46.2
onnx==1.17.0
matplotlib==3.9.2